git-svn-id: file:///Users/braun/svn/vermont/branches/vermont-dynamic-config@1922 aef3b71b-58ee-0310-9ba9-8811b9f0742f

+++ common/Misc.cpp
@@ -1,9 +1,9 @@
-#include <Misc.h>
+#include "Misc.h"
 
 #ifndef _GNU_SOURCE
 size_t strnlen(const char* s, size_t maxlen)
 {
-	int len = 0;
+	size_t len = 0;
 	if (maxlen == 0 || !s) {
 		return 0;
 	}
